Fix black objects in non-Apple environments
Looks like there was a typo leading to division by zero. From the OpenGL
ES spec, "Treatment of conditions such as divide by 0 may lead to an
unspecified result, but must not lead to the interruption or termination
of processing." It's undefined behavior, so each platform gets to define
it themselves! Fun times.